refactor(vue): split workspace preview lifecycle
